+++ apps/rosen/app/_networks/binance/server.ts
@@ -1,20 +1,10 @@
 'use server';
 
-import {
-  generateLockData as generateLockDataCore,
-  generateTxParameters as generateTxParametersCore,
-} from '@rosen-network/evm';
+import { getMaxTransferCreator } from '@rosen-network/binance';
 
 import { wrap } from '@/_safeServerAction';
 import { getTokenMap } from '@/_tokenMap/getServerTokenMap';
 
-export const generateLockData = wrap(generateLockDataCore, {
-  traceKey: 'generateLockData',
+export const getMaxTransfer = wrap(getMaxTransferCreator(getTokenMap()), {
+  traceKey: 'getMaxTransfer',
 });
-
-export const generateTxParameters = wrap(
-  generateTxParametersCore(getTokenMap()),
-  {
-    traceKey: 'generateTxParameters',
-  },
-);