+++ apps/rosen/app/_networks/ethereum/client.ts
@@ -1,33 +1,12 @@
-import { EthereumIcon } from '@rosen-bridge/icons';
-import { NETWORK_LABELS, NETWORKS } from '@rosen-ui/constants';
-import { MetaMaskWallet } from '@rosen-ui/metamask-wallet';
+import { EthereumNetwork } from '@rosen-network/ethereum/dist/client';
 
 import { unwrap } from '@/_safeServerAction';
-import { getTokenMap } from '@/_tokenMap/getClientTokenMap';
-import { EthereumNetwork as EthereumNetworkType } from '@/_types';
 
 import { LOCK_ADDRESSES } from '../../../configs';
-import { getMaxTransfer } from './getMaxTransfer';
-import { generateLockData, generateTxParameters } from './server';
+import { getMaxTransfer } from './server';
 
-/**
- * the main object for Ethereum network
- * providing access to network info and wallets and network specific
- * functionality
- */
-export const EthereumNetwork: EthereumNetworkType = {
-  name: NETWORKS.ETHEREUM,
-  label: NETWORK_LABELS.ETHEREUM,
-  wallets: [
-    new MetaMaskWallet({
-      getTokenMap,
-      generateLockData: unwrap(generateLockData),
-      generateTxParameters: unwrap(generateTxParameters),
-    }),
-  ],
-  logo: EthereumIcon,
-  nextHeightInterval: 50,
+export const ethereumNetwork = new EthereumNetwork({
   lockAddress: LOCK_ADDRESSES.ETHEREUM,
+  nextHeightInterval: 50,
   getMaxTransfer: unwrap(getMaxTransfer),
-  toSafeAddress: (address) => address.toLowerCase(),
-};
+});