+++ apps/rosen/app/_utils/index.ts
@@ -1,3 +1,2 @@
-export * from './getMaxTransfer';
 export * from './getMinTransfer';
 export * from './getTokenNameAndId';